app/requirements.txt
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy>=1.4.0
psycopg2-binary
python-multipart
swisseph>=2.10
websockets>=10.0
python-dotenv
alembic
oauthlib
requests
fastapi>=0.104.0
uvicorn>=0.23.0
strawberry-graphql[fastapi]>=0.215.0